Changes needed because of new releases
* Adding the needed update files
  *	modified:   sql/sigs/CMakeLists.txt
  *	new file:   sql/sigs/pgrouting--3.0.3.sig
  *	new file:   sql/sigs/pgrouting--3.1.1.sig
* updating the release testers
  *	modified:   tools/release-scripts/check-release.sh
  *	modified:   tools/testers/update-tester.sh

+++ tools/testers/update-tester.sh
@@ -26,10 +26,8 @@
 
 set -e
 
-TWEAK=""
 PGPORT=5432
-#sorry this only works on vicky's computer
-# TODO make it more general
+echo "sorry this only works on vicky's computer"
 PGUSER="vicky"
 DB="___pgr___test___"
 
@@ -42,25 +40,30 @@ function info {
 
     echo "EXAMPLE USAGE"
     echo "- Short execution"
-    echo  "bash tools/testers/update-tester.sh $1"
+    echo  "bash tools/testers/update-tester.sh $1 $2"
     echo "- For running pgtap tests:"
-    echo  "bash tools/testers/update-tester.sh $1 long"
+    echo  "bash tools/testers/update-tester.sh $1 $2 long"
 
 }
 
 
 if [[ -z  "$1" ]]; then
     echo missing version example:
-    info 2.6.3
+    info 2.6.3 12
     exit 1
 fi
 
 FROM_PGR="$1"
 CURRENT=$(grep -Po '(?<=project\(PGROUTING VERSION )[^;]+' CMakeLists.txt)
-LONG=$2
+TWEAK=$(grep -Po '(?<=set\(PGROUTING_VERSION_DEV ")[^;]+(?="\))' CMakeLists.txt)
+LONG=$3
+
+if [[ "$FROM_PGR" == 2.* ]];
+then
+    exit 0
+fi
 
 dropdb --if-exists "$DB"
-
 
 cd build
 cmake -DPGROUTING_DEBUG=ON -DCMAKE_BUILD_TYPE=Debug ..
